producer-kafka-consumer-pyspark/requirements.txt
kafka-python
schedule
requests